Consume the reduced review data
The main file will work off of the uncompressed version of
movies.min.csv that you should put in the data folder.

This update will also cache things into Rdata files to make re running
faster.

This code is very memory intensive, but runs to completion on my 8gb
ram mac book

The output file in grapheme is 12 gb

it has yet to be seen if it can be opened in guff